pyproject.toml
[project]
name = "coursework-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Алексей Коновалов",email = "lex-sorokin@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "coursework_1", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
